Ajusts relatório

+++ src/main/java/dc/controller/relatorio/RelatorioFormController.java
@@ -8,13 +8,26 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Controller;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.vaadin.ui.Component;
 
+import dc.controller.financeiro.EmpresaListController;
+import dc.controller.sistema.PapelListController;
+import dc.controller.sistema.SeguimentoListController;
+import dc.controller.sistema.UsuarioListController;
+import dc.entidade.framework.Empresa;
 import dc.entidade.framework.FmMenu;
+import dc.entidade.framework.Papel;
+import dc.entidade.framework.Seguimento;
+import dc.entidade.geral.Usuario;
 import dc.entidade.relatorio.Relatorio;
+import dc.servicos.dao.framework.geral.EmpresaDAO;
 import dc.servicos.dao.framework.geral.FmMenuDAO;
+import dc.servicos.dao.framework.geral.SeguimentoDAO;
 import dc.servicos.dao.relatorio.RelatorioDAO;
+import dc.servicos.dao.sistema.PapelDAO;
+import dc.servicos.dao.sistema.UsuarioDAO;
 import dc.servicos.util.Util;
 import dc.servicos.util.Validator;
 import dc.visao.framework.FmMenuListController;
@@ -41,6 +54,15 @@ public class RelatorioFormController extends CRUDFormController<Relatorio> {
 	@Autowired
 	private FmMenuDAO fmMenuDAO;
 
+	@Autowired
+	private SeguimentoDAO seguimentoDAO;
+	@Autowired
+	private PapelDAO papelDAO;
+	@Autowired
+	private UsuarioDAO usuarioDAO;
+	@Autowired
+	private EmpresaDAO empresaDAO;
+
 	@Override
 	protected String getNome() {
 		return "Relatório";
@@ -52,12 +74,14 @@ public class RelatorioFormController extends CRUDFormController<Relatorio> {
 	}
 
 	@Override
+	@Transactional
 	protected void actionSalvar() {
 		subView.preencheBean(currentBean);
 		try {
 			// currentBean.setEmpresa(SecuritySessionProvider.getUsuario().getConta().getEmpresa());
 			gravarAnexo();
-			relatorioDAO.saveOrUpdate(currentBean);
+			relatorioDAO.salvar(currentBean, subView.getUsuarioContainer().getItemIds(), subView.getSeguimentoContainer().getItemIds(), subView
+					.getPapelContainer().getItemIds(), subView.getEmpresaContainer().getItemIds());
 
 			notifiyFrameworkSaveOK(this.currentBean);
 		} catch (Exception e) {
@@ -66,8 +90,10 @@ public class RelatorioFormController extends CRUDFormController<Relatorio> {
 	}
 
 	@Override
+	@Transactional
 	protected void carregar(Serializable id) {
 		currentBean = relatorioDAO.find(id);
+
 		subView.preencheForm(currentBean);
 	}
 
@@ -92,7 +118,32 @@ public class RelatorioFormController extends CRUDFormController<Relatorio> {
 			}
 		};
 
+		DefaultManyToOneComboModel<Seguimento> seguimentoModel = new DefaultManyToOneComboModel<Seguimento>(SeguimentoListController.class,
+				this.seguimentoDAO, super.getMainController());
+
+		DefaultManyToOneComboModel<Usuario> usuarioModel = new DefaultManyToOneComboModel<Usuario>(UsuarioListController.class, this.usuarioDAO,
+				super.getMainController()) {
+			@Override
+			public String getCaptionProperty() {
+				return "login";
+			}
+		};
+		DefaultManyToOneComboModel<Papel> papelModel = new DefaultManyToOneComboModel<Papel>(PapelListController.class, this.papelDAO,
+				super.getMainController());
+
+		DefaultManyToOneComboModel<Empresa> empresa = new DefaultManyToOneComboModel<Empresa>(EmpresaListController.class, this.empresaDAO,
+				super.getMainController()) {
+			@Override
+			public String getCaptionProperty() {
+				return "nomeFantasia";
+			}
+		};
+
 		subView.getComboMenus().setModel(menusModel);
+		subView.getComboSeguimentos().setModel(seguimentoModel);
+		subView.getComboEmpresas().setModel(empresa);
+		subView.getComboPapeis().setModel(papelModel);
+		subView.getComboUsuarios().setModel(usuarioModel);
 	}
 
 	/*
@@ -135,16 +186,18 @@ public class RelatorioFormController extends CRUDFormController<Relatorio> {
 	private void gravarAnexo() throws IOException {
 		File relatorioForm = (File) subView.getRelatorioUpload().getValue();
 
-		String homePath = System.getProperty("user.home");
-		String customCompanyBaseFolder = "dc-erp/reports";
-		String arqOriginal = subView.getNomeRelatorio();
+		if (relatorioForm != null) {
+			String homePath = System.getProperty("user.home");
+			String customCompanyBaseFolder = "dc-erp/reports";
+			String arqOriginal = subView.getNomeRelatorio();
 
-		String caminho = homePath + "/" + customCompanyBaseFolder + "/" + currentBean.getMenu().getUrlId() + "/" + arqOriginal;
+			String caminho = homePath + "/" + customCompanyBaseFolder + "/" + currentBean.getMenu().getUrlId() + "/" + arqOriginal;
 
-		Util.copyFile(relatorioForm, new File(caminho));
-		currentBean.setJasperPath(caminho);
+			Util.copyFile(relatorioForm, new File(caminho));
+			currentBean.setJasperPath(caminho);
 
-		relatorioForm.delete();
+			relatorioForm.delete();
+		}
 	}
 
 	@Override
